requirements.txt
Flask==0.12.3
psycopg2-binary==2.7.5
pandas==0.23.4
sqlalchemy==1.2.12
beautifulsoup4==4.6.0
sastrawi==1.0.1
scikit-learn==0.19.1
scipy==1.1.0
numpy==1.15.1
APScheduler==3.5.3
newrelic==4.4.1.104







